Remove AGPMInjector
It seems to not be doing anything, at least on AMD GPUs, we have the PowerPlay feature which controls power usage just like on Windows and Linux, from my experience and tests, the only difference it makes is showing up in the I/O Registry Tree. Chances are it silently fails silently to load on our non-Apple VBIOS GPUs, or it just doesn't do anything on modern cards.